feat(banners): renaming spec file to match ECL variant - front-1557
