--- src/ec/packages/ec-component-hero-banner/demo/data--image-text-box.js
@@ -1,6 +1,0 @@
-import specs from '@ecl/ec-specs-hero-banner/demo/data--image';
-import adapter from '../adapter';
-
-specs.variant = 'image';
-
-export default adapter(specs);
