--- src/ec/packages/ec-component-page-banner/demo/data--image-text-box.js
@@ -1,4 +1,0 @@
-import specs from '@ecl/ec-specs-page-banner/demo/data--image';
-import adapter from '../adapter';
-
-export default adapter(specs);
